backend/pyproject.toml
[project]
name = "mlabled-backend"
version = "0.1.0"
description = "MLabled - Image annotation platform backend"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "sqlalchemy[asyncio]>=2.0.0",
    "asyncpg>=0.30.0",
    "alembic>=1.14.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "bcrypt==4.0.1",
    "python-multipart>=0.0.9",
    "boto3>=1.35.0",
    "redis>=5.0.0",
    "celery>=5.4.0",
    "httpx>=0.27.0",
    "pillow>=10.0.0",
]

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"
